Restore methodology source files
